\name{extfunnel-package}
\alias{extfunnel-package}
\docType{package}
\title{
Additional Funnel Plot Augmentations
}
\description{This is a package containing the function \code{extfunnel()} which produces a funnel plot including additional augmentations such as statistical significance contours and heterogeneity contours.}
\details{
\tabular{ll}{
Package: \tab extfunnel\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-29\cr
License: \tab What license is it under?\cr
}
}
\author{
Dean Langan (author and maintainer): <dl790@york.ac.uk>

}

\keyword{ package }
\keyword{Meta-analysis}
\keyword{Funnel}
\keyword{plot}
\keyword{graph}